requirements.txt
# python == 3.8.18
gensim==4.3.2
h5py
matplotlib
numpy
pandas
Pillow
scikit-image
scikit-learn
scipy==1.9.1
seaborn
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm
